Added dataproc (pyspark) code with deployment script. It currently uses workflow templates for deployment with a managed cluster that gets automatically deleted after the job(s) complete.

+++ data_pipelines/batch/transactions_etl_to_dwh/transactions_etl/transforms.py
@@ -53,7 +53,7 @@ def enrich_with_dimensions(transactions_df: DataFrame, dims: dict) -> DataFrame:
     # Prepare date and time keys for joining
     df_with_join_keys = transactions_df \
         .withColumn("Date_Join_Key", date_format(to_date(col("timestamp")), "yyyyMMdd").cast("integer")) \
-        .withColumn("Time_Join_Key", date_format(col("timestamp"), "HHmmss"))
+        .withColumn("Time_Join_Key", date_format(col("timestamp"), "HHmmss").cast("integer"))
 
     # Join with dimension tables. Using broadcast for smaller tables is a key optimization.
     enriched_df = df_with_join_keys \
